[Alerting] Refactored External Service Simulators (#43570)
