+++ x-pack/test/alerting_api_integration/common/config.ts
@@ -8,7 +8,10 @@ import path from 'path';
 import { CA_CERT_PATH } from '@kbn/dev-utils';
 import { FtrConfigProviderContext } from '@kbn/test/types/ftr';
 import { services } from './services';
-import { SLACK_ACTION_SIMULATOR_URI } from './fixtures/plugins/actions';
+import {
+  getExternalServiceSimulatorPath,
+  ExternalServiceSimulator,
+} from './fixtures/plugins/actions';
 
 interface CreateTestConfigOptions {
   license: string;
@@ -56,7 +59,9 @@ export function createTestConfig(name: string, options: CreateTestConfigOptions)
           ...disabledPlugins.map(key => `--xpack.${key}.enabled=false`),
           `--plugin-path=${path.join(__dirname, 'fixtures', 'plugins', 'alerts')}`,
           `--plugin-path=${path.join(__dirname, 'fixtures', 'plugins', 'actions')}`,
-          `--server.xsrf.whitelist=${JSON.stringify([SLACK_ACTION_SIMULATOR_URI])}`,
+          `--server.xsrf.whitelist=${JSON.stringify([
+            getExternalServiceSimulatorPath(ExternalServiceSimulator.SLACK),
+          ])}`,
           ...(ssl
             ? [
                 `--elasticsearch.hosts=${servers.elasticsearch.protocol}://${servers.elasticsearch.hostname}:${servers.elasticsearch.port}`,
