+++ x-pack/test/alerting_api_integration/common/fixtures/plugins/actions/index.ts
@@ -3,18 +3,17 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-
-import Joi from 'joi';
 import Hapi from 'hapi';
+import { initPlugin as initSlack } from './slack_simulation';
 
 const NAME = 'actions-FTS-external-service-simulators';
 
-export const SLACK_ACTION_SIMULATOR_URI = `/api/_${NAME}/slack`;
+export enum ExternalServiceSimulator {
+  SLACK = 'slack',
+}
 
-interface SlackRequest extends Hapi.Request {
-  payload: {
-    text: string;
-  };
+export function getExternalServiceSimulatorPath(service: ExternalServiceSimulator): string {
+  return `/api/_${NAME}/${service}`;
 }
 
 // eslint-disable-next-line import/no-default-export
@@ -22,75 +21,8 @@ export default function(kibana: any) {
   return new kibana.Plugin({
     require: ['actions'],
     name: NAME,
-    init: initPlugin,
-  });
-}
-
-function initPlugin(server: any) {
-  server.route({
-    method: 'POST',
-    path: `${SLACK_ACTION_SIMULATOR_URI}`,
-    options: {
-      auth: false,
-      validate: {
-        options: { abortEarly: false },
-        payload: Joi.object().keys({
-          text: Joi.string(),
-        }),
-      },
+    init: (server: Hapi.Server) => {
+      initSlack(server, getExternalServiceSimulatorPath(ExternalServiceSimulator.SLACK));
     },
-    handler: slackHandler,
   });
-}
-
-// Slack simulator: create a slack action pointing here, and you can get
-// different responses based on the message posted. See the README.md for
-// more info.
-
-function slackHandler(request: SlackRequest, h: any) {
-  const body = request.payload;
-  const text = body && body.text;
-
-  if (text == null) {
-    return htmlResponse(h, 400, 'bad request to slack simulator');
-  }
-
-  switch (text) {
-    case 'success':
-      return htmlResponse(h, 200, 'ok');
-
-    case 'no_text':
-      return htmlResponse(h, 400, 'no_text');
-
-    case 'invalid_payload':
-      return htmlResponse(h, 400, 'invalid_payload');
-
-    case 'invalid_token':
-      return htmlResponse(h, 403, 'invalid_token');
-
-    case 'status_500':
-      return htmlResponse(h, 500, 'simulated slack 500 response');
-
-    case 'rate_limit':
-      const response = {
-        retry_after: 1,
-        ok: false,
-        error: 'rate_limited',
-      };
-
-      return h
-        .response(response)
-        .type('application/json')
-        .header('retry-after', '1')
-        .code(429);
-  }
-
-  return htmlResponse(h, 400, 'unknown request to slack simulator');
-}
-
-function htmlResponse(h: any, code: number, text: string) {
-  return h
-    .response(text)
-    .type('text/html')
-    .code(code);
 }
